adds AccountName component
